gvle: add a VleCodeEdit widget
Just moving the one available inside tge default cpp editor
to the collection of vle widgets.
The default cpp editor does use this moved widget.
