+++ src/pkgs/gvle.default/DefaultCppPanel/DefaultCppPanel.cpp
@@ -24,265 +24,16 @@
 
 #include "DefaultCppPanel.h"
 #include <QFileInfo>
-#include <QFontDatabase>
-#include <QPainter>
 #include <sstream>
 #include <vle/utils/Template.hpp>
 
 namespace vle {
 namespace gvle {
 
-Highlighter::Highlighter(QTextDocument* parent)
-  : QSyntaxHighlighter(parent)
-{
-    HighlightingRule rule;
-
-    keywordFormat.setForeground(Qt::darkBlue);
-    keywordFormat.setFontWeight(QFont::Bold);
-    QStringList keywordPatterns;
-    keywordPatterns << "\\bchar\\b"
-                    << "\\bclass\\b"
-                    << "\\bconst\\b"
-                    << "\\bdouble\\b"
-                    << "\\benum\\b"
-                    << "\\bexplicit\\b"
-                    << "\\bfriend\\b"
-                    << "\\binline\\b"
-                    << "\\bint\\b"
-                    << "\\blong\\b"
-                    << "\\bnamespace\\b"
-                    << "\\boperator\\b"
-                    << "\\bprivate\\b"
-                    << "\\bprotected\\b"
-                    << "\\bpublic\\b"
-                    << "\\bshort\\b"
-                    << "\\bsignals\\b"
-                    << "\\bsigned\\b"
-                    << "\\bslots\\b"
-                    << "\\bstatic\\b"
-                    << "\\bstruct\\b"
-                    << "\\btemplate\\b"
-                    << "\\btypedef\\b"
-                    << "\\btypename\\b"
-                    << "\\bunion\\b"
-                    << "\\bunsigned\\b"
-                    << "\\bvirtual\\b"
-                    << "\\bvoid\\b"
-                    << "\\bvolatile\\b";
-    foreach (const QString& pattern, keywordPatterns) {
-        rule.pattern = QRegExp(pattern);
-        rule.format = keywordFormat;
-        highlightingRules.append(rule);
-    }
-
-    classFormat.setFontWeight(QFont::Bold);
-    classFormat.setForeground(Qt::darkMagenta);
-    rule.pattern = QRegExp("\\bQ[A-Za-z]+\\b");
-    rule.format = classFormat;
-    highlightingRules.append(rule);
-
-    quotationFormat.setForeground(Qt::darkGreen);
-    rule.pattern = QRegExp("\".*\"");
-    rule.format = quotationFormat;
-    highlightingRules.append(rule);
-
-    functionFormat.setFontItalic(true);
-    functionFormat.setForeground(Qt::blue);
-    rule.pattern = QRegExp("\\b[A-Za-z0-9_]+(?=\\()");
-    rule.format = functionFormat;
-    highlightingRules.append(rule);
-
-    singleLineCommentFormat.setForeground(Qt::red);
-    rule.pattern = QRegExp("//[^\n]*");
-    rule.format = singleLineCommentFormat;
-    highlightingRules.append(rule);
-
-    multiLineCommentFormat.setForeground(Qt::red);
-
-    commentStartExpression = QRegExp("/\\*");
-    commentEndExpression = QRegExp("\\*/");
-}
-
-void
-Highlighter::highlightBlock(const QString& text)
-{
-    foreach (const HighlightingRule& rule, highlightingRules) {
-        QRegExp expression(rule.pattern);
-        int index = expression.indexIn(text);
-        while (index >= 0) {
-            int length = expression.matchedLength();
-            setFormat(index, length, rule.format);
-            index = expression.indexIn(text, index + length);
-        }
-    }
-
-    setCurrentBlockState(0);
-
-    int startIndex = 0;
-    if (previousBlockState() != 1)
-        startIndex = commentStartExpression.indexIn(text);
-
-    while (startIndex >= 0) {
-        int endIndex = commentEndExpression.indexIn(text, startIndex);
-        int commentLength;
-        if (endIndex == -1) {
-            setCurrentBlockState(1);
-            commentLength = text.length() - startIndex;
-        } else {
-            commentLength =
-              endIndex - startIndex + commentEndExpression.matchedLength();
-        }
-        setFormat(startIndex, commentLength, multiLineCommentFormat);
-        startIndex =
-          commentStartExpression.indexIn(text, startIndex + commentLength);
-    }
-}
-
-LineNumberArea::LineNumberArea(CodeEditor* editor)
-  : QWidget(editor)
-  , codeEditor(editor)
-{
-}
-
-QSize
-LineNumberArea::sizeHint() const
-{
-    return QSize(codeEditor->lineNumberAreaWidth(), 0);
-}
-
-void
-LineNumberArea::paintEvent(QPaintEvent* event)
-{
-    codeEditor->lineNumberAreaPaintEvent(event);
-}
-
-CodeEditor::CodeEditor(QWidget* parent)
-  : QPlainTextEdit(parent)
-{
-    QFont font = QFontDatabase::systemFont(QFontDatabase::FixedFont);
-    font.setPointSize(9);
-    setFont(font);
-
-    m_highlighter = new Highlighter(document());
-    m_lineNumberArea = new LineNumberArea(this);
-
-    connect(this,
-            SIGNAL(blockCountChanged(int)),
-            this,
-            SLOT(updateLineNumberAreaWidth(int)));
-    connect(this,
-            SIGNAL(updateRequest(QRect, int)),
-            this,
-            SLOT(updateLineNumberArea(QRect, int)));
-    connect(this,
-            SIGNAL(cursorPositionChanged()),
-            this,
-            SLOT(highlightCurrentLine()));
-
-    updateLineNumberAreaWidth(0);
-    highlightCurrentLine();
-
-    //    QObject::connect(m_edit, SIGNAL(undoAvailable(bool)),
-    //                     this, SLOT(onUndoAvailable(bool)));
-}
-
-int
-CodeEditor::lineNumberAreaWidth()
-{
-    int digits = 1;
-    int max = qMax(1, blockCount());
-    while (max >= 10) {
-        max /= 10;
-        ++digits;
-    }
-
-    int space = 3 + fontMetrics().width(QLatin1Char('9')) * digits;
-
-    return space;
-}
-
-void
-CodeEditor::updateLineNumberAreaWidth(int /* newBlockCount */)
-{
-    setViewportMargins(lineNumberAreaWidth(), 0, 0, 0);
-}
-
-void
-CodeEditor::updateLineNumberArea(const QRect& rect, int dy)
-{
-    if (dy)
-        m_lineNumberArea->scroll(0, dy);
-    else
-        m_lineNumberArea->update(
-          0, rect.y(), m_lineNumberArea->width(), rect.height());
-
-    if (rect.contains(viewport()->rect()))
-        updateLineNumberAreaWidth(0);
-}
-
-void
-CodeEditor::resizeEvent(QResizeEvent* e)
-{
-    QPlainTextEdit::resizeEvent(e);
-
-    QRect cr = contentsRect();
-    m_lineNumberArea->setGeometry(
-      QRect(cr.left(), cr.top(), lineNumberAreaWidth(), cr.height()));
-}
-
-void
-CodeEditor::highlightCurrentLine()
-{
-    QList<QTextEdit::ExtraSelection> extraSelections;
-
-    if (!isReadOnly()) {
-        QTextEdit::ExtraSelection selection;
-
-        QColor lineColor = QColor(Qt::yellow).lighter(160);
-
-        selection.format.setBackground(lineColor);
-        selection.format.setProperty(QTextFormat::FullWidthSelection, true);
-        selection.cursor = textCursor();
-        selection.cursor.clearSelection();
-        extraSelections.append(selection);
-    }
-
-    setExtraSelections(extraSelections);
-}
-
-void
-CodeEditor::lineNumberAreaPaintEvent(QPaintEvent* event)
-{
-    QPainter painter(m_lineNumberArea);
-    painter.fillRect(event->rect(), Qt::lightGray);
-    QTextBlock block = firstVisibleBlock();
-    int blockNumber = block.blockNumber();
-    int top =
-      (int)blockBoundingGeometry(block).translated(contentOffset()).top();
-    int bottom = top + (int)blockBoundingRect(block).height();
-    while (block.isValid() && top <= event->rect().bottom()) {
-        if (block.isVisible() && bottom >= event->rect().top()) {
-            QString number = QString::number(blockNumber + 1);
-            painter.setPen(Qt::black);
-            painter.drawText(0,
-                             top,
-                             m_lineNumberArea->width(),
-                             fontMetrics().height(),
-                             Qt::AlignRight,
-                             number);
-        }
-
-        block = block.next();
-        top = bottom;
-        bottom = top + (int)blockBoundingRect(block).height();
-        ++blockNumber;
-    }
-}
-
 DefaultCppPanel::DefaultCppPanel()
   : PluginMainPanel()
 {
-    m_edit = new CodeEditor;
+    m_edit = new VleCodeEdit;
 }
 
 DefaultCppPanel::~DefaultCppPanel()
