+++ src/vle/gvle/gvle_widgets.cpp
@@ -34,6 +34,7 @@
 #include <QScrollBar>
 #include <QTableWidget>
 #include <QtDebug>
+#include <QPainter>
 ////#include <QtWidgets/QStackedWidget>
 //#include <QStackedWidget>
 
@@ -1797,5 +1798,369 @@ VleValueWidget::setComplexWidget(const QString& id,
     table->setCellWidget(r, c, w);
     table->setItem(r, c, new QTableWidgetItem);
 }
+
+VleCodeHighlighter::VleCodeHighlighter(QTextDocument* parent)
+  : QSyntaxHighlighter(parent)
+{
+    HighlightingRule rule;
+
+    keywordFormat.setForeground(Qt::darkBlue);
+    keywordFormat.setFontWeight(QFont::Bold);
+    QStringList keywordPatterns;
+    keywordPatterns << "\\bchar\\b"
+                    << "\\bclass\\b"
+                    << "\\bconst\\b"
+                    << "\\bdouble\\b"
+                    << "\\benum\\b"
+                    << "\\bexplicit\\b"
+                    << "\\bfriend\\b"
+                    << "\\binline\\b"
+                    << "\\bint\\b"
+                    << "\\blong\\b"
+                    << "\\bnamespace\\b"
+                    << "\\boperator\\b"
+                    << "\\bprivate\\b"
+                    << "\\bprotected\\b"
+                    << "\\bpublic\\b"
+                    << "\\bshort\\b"
+                    << "\\bsignals\\b"
+                    << "\\bsigned\\b"
+                    << "\\bslots\\b"
+                    << "\\bstatic\\b"
+                    << "\\bstruct\\b"
+                    << "\\btemplate\\b"
+                    << "\\btypedef\\b"
+                    << "\\btypename\\b"
+                    << "\\bunion\\b"
+                    << "\\bunsigned\\b"
+                    << "\\bvirtual\\b"
+                    << "\\bvoid\\b"
+                    << "\\bvolatile\\b";
+    foreach (const QString& pattern, keywordPatterns) {
+        rule.pattern = QRegExp(pattern);
+        rule.format = keywordFormat;
+        highlightingRules.append(rule);
+    }
+
+    classFormat.setFontWeight(QFont::Bold);
+    classFormat.setForeground(Qt::darkMagenta);
+    rule.pattern = QRegExp("\\bQ[A-Za-z]+\\b");
+    rule.format = classFormat;
+    highlightingRules.append(rule);
+
+    quotationFormat.setForeground(Qt::darkGreen);
+    rule.pattern = QRegExp("\".*\"");
+    rule.format = quotationFormat;
+    highlightingRules.append(rule);
+
+    functionFormat.setFontItalic(true);
+    functionFormat.setForeground(Qt::blue);
+    rule.pattern = QRegExp("\\b[A-Za-z0-9_]+(?=\\()");
+    rule.format = functionFormat;
+    highlightingRules.append(rule);
+
+    singleLineCommentFormat.setForeground(Qt::red);
+    rule.pattern = QRegExp("//[^\n]*");
+    rule.format = singleLineCommentFormat;
+    highlightingRules.append(rule);
+
+    multiLineCommentFormat.setForeground(Qt::red);
+
+    commentStartExpression = QRegExp("/\\*");
+    commentEndExpression = QRegExp("\\*/");
+}
+
+void
+VleCodeHighlighter::highlightBlock(const QString& text)
+{
+    foreach (const HighlightingRule& rule, highlightingRules) {
+        QRegExp expression(rule.pattern);
+        int index = expression.indexIn(text);
+        while (index >= 0) {
+            int length = expression.matchedLength();
+            setFormat(index, length, rule.format);
+            index = expression.indexIn(text, index + length);
+        }
+    }
+
+    setCurrentBlockState(0);
+
+    int startIndex = 0;
+    if (previousBlockState() != 1)
+        startIndex = commentStartExpression.indexIn(text);
+
+    while (startIndex >= 0) {
+        int endIndex = commentEndExpression.indexIn(text, startIndex);
+        int commentLength;
+        if (endIndex == -1) {
+            setCurrentBlockState(1);
+            commentLength = text.length() - startIndex;
+        } else {
+            commentLength =
+              endIndex - startIndex + commentEndExpression.matchedLength();
+        }
+        setFormat(startIndex, commentLength, multiLineCommentFormat);
+        startIndex =
+          commentStartExpression.indexIn(text, startIndex + commentLength);
+    }
+}
+
+VleCodeLineNumberArea::VleCodeLineNumberArea(VleCodeEdit* editor)
+  : QWidget(editor)
+  , codeEditor(editor)
+{
+}
+
+QSize
+VleCodeLineNumberArea::sizeHint() const
+{
+    return QSize(codeEditor->VleCodelineNumberAreaWidth(), 0);
+}
+
+void
+VleCodeLineNumberArea::paintEvent(QPaintEvent* event)
+{
+    codeEditor->VleCodelineNumberAreaPaintEvent(event);
+}
+
+VleCodeEdit::VleCodeEdit(QWidget* parent)
+  : QPlainTextEdit(parent)
+{
+    QFont font = QFontDatabase::systemFont(QFontDatabase::FixedFont);
+    font.setPointSize(9);
+    setFont(font);
+
+    m_highlighter = new VleCodeHighlighter(document());
+    m_lineNumberArea = new VleCodeLineNumberArea(this);
+
+    connect(this,
+            SIGNAL(blockCountChanged(int)),
+            this,
+            SLOT(updateVleCodeLineNumberAreaWidth(int)));
+    connect(this,
+            SIGNAL(updateRequest(QRect, int)),
+            this,
+            SLOT(updateVleCodeLineNumberArea(QRect, int)));
+    connect(this,
+            SIGNAL(cursorPositionChanged()),
+            this,
+            SLOT(highlightCurrentLine()));
+
+    updateVleCodeLineNumberAreaWidth(0);
+    highlightCurrentLine();
+
+    //    QObject::connect(m_edit, SIGNAL(undoAvailable(bool)),
+    //                     this, SLOT(onUndoAvailable(bool)));
+}
+
+VleCodeEdit::VleCodeEdit(QWidget* parent,
+                         const QString& text,
+                         const QString& idstr,
+                         bool editOnDbleClick)
+  : QPlainTextEdit(parent)
+  , saved_value("")
+  , id(idstr)
+  , edit_on_dble_click(editOnDbleClick)
+{
+    QFont font = QFontDatabase::systemFont(QFontDatabase::FixedFont);
+    font.setPointSize(9);
+    setFont(font);
+
+    m_highlighter = new VleCodeHighlighter(document());
+    m_lineNumberArea = new VleCodeLineNumberArea(this);
+
+    connect(this,
+            SIGNAL(blockCountChanged(int)),
+            this,
+            SLOT(updateVleCodeLineNumberAreaWidth(int)));
+    connect(this,
+            SIGNAL(updateRequest(QRect, int)),
+            this,
+            SLOT(updateVleCodeLineNumberArea(QRect, int)));
+    connect(this,
+            SIGNAL(cursorPositionChanged()),
+            this,
+            SLOT(highlightCurrentLine()));
+
+    updateVleCodeLineNumberAreaWidth(0);
+    highlightCurrentLine();
+
+    this->setLineWrapMode(QPlainTextEdit::NoWrap);
+    // this->setHorizontalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
+    // this->setVerticalScrollBarPolicy(Qt::ScrollBarAlwaysOff);
+    // this->setTextInteractionFlags(Qt::NoTextInteraction);
+    // this->setContextMenuPolicy(Qt::NoContextMenu);
+    this->setFocusPolicy(Qt::ClickFocus);
+
+    const int tabStop = 4;  // 4 characters
+
+    QFontMetrics metrics(font);
+    this->setTabStopWidth(tabStop * metrics.width(' '));
+
+    setText(text);
+    this->resize(document()->size().toSize());
+}
+
+void
+VleCodeEdit::setTextEdition(bool val)
+{
+
+    if (val and this->textInteractionFlags() == Qt::NoTextInteraction) {
+        saved_value = document()->toPlainText();
+        this->setFocus(
+          Qt::ActiveWindowFocusReason); // this gives the item keyboard focus
+        this->setTextInteractionFlags(Qt::TextEditorInteraction);
+    } else {
+        this->setTextInteractionFlags(Qt::NoTextInteraction);
+        this->clearFocus();
+        if (document()->toPlainText() != saved_value) {
+            emit textUpdated(id, saved_value, document()->toPlainText());
+        }
+    }
+}
+
+QString
+VleCodeEdit::getSavedText()
+{
+    return saved_value;
+}
+QString
+VleCodeEdit::getCurrentText()
+{
+    return document()->toPlainText();
+}
+
+void
+VleCodeEdit::setText(const QString& text)
+{
+    this->document()->setPlainText(text);
+    saved_value = text;
+}
+
+void
+VleCodeEdit::focusOutEvent(QFocusEvent* e)
+{
+    //setPalette(QApplication::palette());
+    setTextEdition(false);
+    update();
+    QPlainTextEdit::focusOutEvent(e);
+}
+
+void
+VleCodeEdit::focusInEvent(QFocusEvent* e)
+{
+    if (e->reason() == Qt::MouseFocusReason) {
+        //QPalette p = this->palette();
+        //p.setBrush(QPalette::Base,
+        //           QBrush(QColor(255, 127, 80, 200), Qt::SolidPattern));
+    //setPalette(p);
+        QPlainTextEdit::focusInEvent(e);
+        emit selected(id);
+    }
+}
+
+void
+VleCodeEdit::mouseDoubleClickEvent(QMouseEvent* e)
+{
+    if (edit_on_dble_click) {
+        setTextEdition(true);
+    } else {
+        QPlainTextEdit::mouseDoubleClickEvent(e);
+    }
+}
+
+int
+VleCodeEdit::VleCodelineNumberAreaWidth()
+{
+    int digits = 1;
+    int max = qMax(1, blockCount());
+    while (max >= 10) {
+        max /= 10;
+        ++digits;
+    }
+
+    int space = 3 + fontMetrics().width(QLatin1Char('9')) * digits;
+
+    return space;
+}
+
+void
+VleCodeEdit::updateVleCodeLineNumberAreaWidth(int /* newBlockCount */)
+{
+    setViewportMargins(VleCodelineNumberAreaWidth(), 0, 0, 0);
+}
+
+void
+VleCodeEdit::updateVleCodeLineNumberArea(const QRect& rect, int dy)
+{
+    if (dy)
+        m_lineNumberArea->scroll(0, dy);
+    else
+        m_lineNumberArea->update(
+          0, rect.y(), m_lineNumberArea->width(), rect.height());
+
+    if (rect.contains(viewport()->rect()))
+        updateVleCodeLineNumberAreaWidth(0);
+}
+
+void
+VleCodeEdit::resizeEvent(QResizeEvent* e)
+{
+    QPlainTextEdit::resizeEvent(e);
+
+    QRect cr = contentsRect();
+    m_lineNumberArea->setGeometry(
+      QRect(cr.left(), cr.top(), VleCodelineNumberAreaWidth(), cr.height()));
+}
+
+void
+VleCodeEdit::highlightCurrentLine()
+{
+    QList<QTextEdit::ExtraSelection> extraSelections;
+
+    if (!isReadOnly()) {
+        QTextEdit::ExtraSelection selection;
+
+        QColor lineColor = QColor(Qt::yellow).lighter(160);
+
+        selection.format.setBackground(lineColor);
+        selection.format.setProperty(QTextFormat::FullWidthSelection, true);
+        selection.cursor = textCursor();
+        selection.cursor.clearSelection();
+        extraSelections.append(selection);
+    }
+
+    setExtraSelections(extraSelections);
+}
+
+void
+VleCodeEdit::VleCodelineNumberAreaPaintEvent(QPaintEvent* event)
+{
+    QPainter painter(m_lineNumberArea);
+    painter.fillRect(event->rect(), Qt::lightGray);
+    QTextBlock block = firstVisibleBlock();
+    int blockNumber = block.blockNumber();
+    int top =
+      (int)blockBoundingGeometry(block).translated(contentOffset()).top();
+    int bottom = top + (int)blockBoundingRect(block).height();
+    while (block.isValid() && top <= event->rect().bottom()) {
+        if (block.isVisible() && bottom >= event->rect().top()) {
+            QString number = QString::number(blockNumber + 1);
+            painter.setPen(Qt::black);
+            painter.drawText(0,
+                             top,
+                             m_lineNumberArea->width(),
+                             fontMetrics().height(),
+                             Qt::AlignRight,
+                             number);
+        }
+
+        block = block.next();
+        top = bottom;
+        bottom = top + (int)blockBoundingRect(block).height();
+        ++blockNumber;
+    }
+}
+
 }
 } // namespaces
